requirements.txt
Flask
opencv-python
numpy
python-docx
pytest
python-docx